Add Python 3.14 testing and wheel building
This change adds 3.14 to the lists of Python versions to test and to the
list to build with cibuildwheel. It also marks the package as compatible
with Python 3.13 and 3.14. It does not drop Python 3.9 or update the
Python versions used in parts of the GitHub workflow that only use a
single Python version, though doing those updates would also be a good
idea.